backend-fast-api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
limits==4.0.1
packaging==24.2
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
motor==3.7.0
pymongo==4.11.2
